ai/requirements.txt
fastapi
uvicorn
mangum
pydantic-settings
boto3
httpx
python-multipart
python-dotenv
beautifulsoup4
PyMuPDF==1.25.3
openai
pyproject_hooks
build
wheel
setuptools
pip
chromadb
sentence-transformers
numpy<2.0.0
rank-bm25>=0.2.2
PyJWT
protobuf==3.20.3
